ZKVM-946: Add field extension operations (#2728)
Add field extension addition, subtraction, and multiplication bigint
precompiles. There are general forms that take the (non-leading)
coefficients of the monic irreducible polynomial of the extensions for
degree 2 extensions. (And, for multiplication, for a degree 4
extension.) Also, for multiplication, there's a specialized version
that's more efficient specifically for the extension F[x]/(x^2 + 1).

Closes ZKVM-912

---------

Co-authored-by: iddo <iddo@risczero.com>